First steps toward ingesting books as datastore objects
* Improvements to creation of `books.json` index file during build
* Application route retrieves and parses `books.json`
* First pass at defining models for `book` and `section`

*Next steps:* for each entry in `books.json`, app route should confirm
that there is a corresponding DB entry, and create on if there isn't.
Bonus points if we also update DB entries if they exist but differ from
the content in `books.json`.

The `books.json` file is derived from book-specific metadata in
`public/books/NAME_OF_BOOK/book.json` (that directory is also expected
to contain the book's HTML content). An example of the schema being used
for this commit follows:

```
 cat public/books/CSE/book.json
{
    "title": "Common Sense Economics",
    "publish_date": 20160101,
    "publisher": "Some folks",
    "cover_image": null,
    "category": "economics",
    "sections": [
        {
            "title": "The whole book (for now)",
            "start_page": "book.html"
        }
    ],
    "schema_version": 2016041800
}
```

For now, if you want to test just create a file like that, run `ember build`, and look at the browser console output when you load the app, and you should see evidence that it has been accessed by the app, though as of this commit nothing is stored to the database.

If you put HTML into the CSE directory then you should also be able to see that content by going to e.g. http://localhost:4200/books/CSE/book.html (should probably make this something like `/content/books/...` or something to avoid confusion with the `/book/...` route defined in Ember).