Click registered, statement stored, but not accepted by LRS
I think I've got a solution for registering clicks and other actions
that will make it easy to add this functionality wherever we need it.

The code that records xAPI statements is stored in the root route. There
is one method, `recordxAPI(statement_data)`, which adds all the common
boilerplate (`actor`, (eventually) `authority`, etc). It is called from
actions like `xAPIOpenLink(event)`, which fills in the event-dependent
parts of the statement before handing off to `recordxAPI`.

My making the main interface an action, we gain two things:

1. Access to the `store` object, which isn't normally injected into
components (and the way you cause it to be injected looks rather ugly
IMO).

2. Because of the way Ember handles actions, you can call e.g.
`xAPIOpenLink` from anywhere, and it will eventually bubble up to the
root. In the case of components, you just have to be sure to add
something like `onOpenLink = 'onxAPIOpenLink'` in the template
(***Note:*** closure-style actions don't bubble, so it has to use the
older syntax), and then have something in the component's code that calls
`this.sendAction('onOpenLink')`.

***TODO:***
The statement is being stored in the database, but there are (at least) two problems:

* The user information is undefined, despite `currentUser` being injected
into the route.
* The LRS rejects the statement, and probably would even if the user stuff
was correct, because there is no `authority` section. Need to get this info from the settings and include it.