WIP: Initial attempt at persisting reading location
The pieces that should be needed are all in place, and everything
builds, but attempting to open the book results in a bunch of
`malformed urf-8` errors.

I probably won't have time to look at this for a while, so I'm checking
it in as-is in the hope that someone else will have a chance to review
and fix.