Fix for broken next/prev buttons
Making this block conditional on the `scrolling` property being false breaks navigation, because the property is only set to false when `didScroll` is run, which is what this block does.

In other words, the conditional that this revision removes keeps `scrolling` stuck set to `true`, which prevents the code that actually advances the text from running.